src/requirements.txt
boto3==1.35.76
botocore==1.35.76
Pillow==11.0.0
